feat(pwa): silent auto-update, changelog as sole "what's new" surface
The "Mise à jour disponible" toast duplicated the changelog dialog as a
second "new version" surface. Drop the toast and let the changelog be the
single what's-new surface.

- Switch registerType from 'prompt' to 'autoUpdate' so new builds activate
  and reload silently in the background; the ChangelogDialog then surfaces
  the new version once the fresh bundle loads.
- sw.ts now skipWaiting + clients.claim on activate (injectManifest doesn't
  inject these for us) and drops the now-unused SKIP_WAITING message branch.
- Replace PWAUpdatePrompt with a headless PWALifecycle bridge that keeps the
  offline-ready toast and the SET_LOCALE service-worker locale sync.
- Remove the dead pwa.updatePanel.* translations (fr + en).
